requirements.txt
ipykernel
ipython
ipython_genutils
xlrd
gensim
scipy
seaborn
spacy
nltk
scikit-learn
thefuzz
typing
umap-learn
transformers
torch
torchaudio
torchvision
urllib3<2.0
langchain
langchain-openai